requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.4.0
bitarray==3.8.0
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
ckzg==2.1.5
click==8.3.1
colorama==0.4.6
cryptography==41.0.1
cytoolz==1.1.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.0.1
Flask==2.3.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
frozenlist==1.8.0
greenlet==3.3.0
hexbytes==1.3.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lru-dict==1.4.1
MarkupSafe==3.0.3
multidict==6.7.0
parsimonious==0.9.0
propcache==0.4.1
protobuf==6.33.2
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
pyunormalize==17.0.0
pywin32==311
referencing==0.37.0
regex==2025.11.3
requests==2.31.0
rlp==4.1.0
rpds-py==0.30.0
setuptools==80.9.0
SQLAlchemy==2.0.44
toolz==1.1.0
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.0
web3==7.14.0
websockets==15.0.1
Werkzeug==2.3.6
yarl==1.22.0
